refactor: extract usePositionSummary composable (#185)
The PositionSummaryWidget script block had grown to ~390 lines,
forcing position-display logic to live alongside the template. Moving
the computeds and tx handlers into a co-located composable keeps the
template a thin projection of already-computed values, matching the
house Vue style and the CloseDialog/RepayDialog splits in this epic.

Logic is moved verbatim — the SHORT/LONG size-rounding branches and the
status switch are byte-for-byte unchanged, so the rounding regression
net stays valid. Two inline navigation handlers become named functions.
